requirements.txt
ccxt
pandas
numpy
ta-lib
python-dotenv
requests